backend/requirements.txt
Flask>=2.2.0
Flask-CORS>=3.0.10
Flask-SQLAlchemy>=3.0.2
Flask-Migrate>=4.0.0
psycopg2-binary>=2.9.1
python-dotenv>=0.19.0
SQLAlchemy>=2.0.0
Werkzeug>=2.2.0
email-validator>=2.0.0
gunicorn>=20.1.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=6.0.0
